+++ db/migrate/20210405140625_add_score_and_completed_at_to_drives_candidate.rb
